common/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.7.4
pydantic-settings==2.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
cryptography==41.0.7
python-dotenv==1.0.0
email-validator==2.3.0
Pillow==10.1.0
slowapi==0.1.9
